refactor: update investors page
