+++ src/templates/Investors/Advantages/Protocol/styles.ts
@@ -1,68 +1,5 @@
 import styled from 'styled-components'
 
-export const Products = styled.section`
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 30.7rem;
-  padding: 0 2.4rem;
-
-  @media (max-width: 992px) {
-    padding-right: 2.4rem;
-    padding-left: 2.4rem;
-    margin-top: 7.6rem;
-  }
-
-  @media (max-width: 576px) {
-    padding-right: 1.6rem;
-    padding-left: 1.6rem;
-  }
-`
-
-export const TitleWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  > h2 {
-    text-align: center;
-  }
-
-  > p {
-    margin-top: 2.4rem;
-  }
-
-  @media (max-width: 576px) {
-    h1 {
-      font-size: 2.4rem;
-      line-height: 3.2rem;
-      font-weight: 700;
-    }
-
-    span {
-      font-size: 1.6rem;
-      line-height: 2.4rem;
-      font-weight: 300;
-    }
-  }
-`
-
-export const FundsCards = styled.div`
-  width: 100%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 5.6rem;
-  gap: 3.65rem;
-
-  @media (max-width: 690px) {
-    flex-direction: column;
-  }
-`
-
 export const Protocol = styled.div`
   width: 100%;
